Fix import operators being broken in version 2.93
Member filename_ext is not a property so should not have been changed to
the declaration with ":". Not sure why this was not a problem in earlier
versions, but it definitely breaks in 2.93. Perhaps due to the new
python version. Anyways, fix is simply rolling back to normal "=" for
filename_ext. Also fix typo in ImportPopWows operator for that member.